+++ resources/themes/pterodactyl/admin/users/view.blade.php
@@ -126,7 +126,7 @@
                         </tr>
                     </thead>
                     <tbody>
-                        @foreach($user->access()->get() as $server)
+                        @foreach($user->setAccessLevel('subuser')->access()->get() as $server)
                             <tr>
                                 <td><a href="{{ route('server.index', $server->uuidShort) }}/"><i class="fa fa-tachometer"></i></a></td>
                                 <td><code>{{ $server->uuidShort }}</code></td>