+++ CHANGELOG.md
@@ -22,6 +22,7 @@ This project follows [Semantic Versioning](http://semver.org) guidelines.
 * Mobile views are now more... viewable. Fixes `col-xs-6` usage thoughout the Admin CP where it was intended to be `col-md-6`.
 * Node Configuration tokens and Download tokens are stored using the cache helpers rather than a database to speed up functions and make use of auto-expiration/deletion functions.
 * Old daemon routes using `/remote` have been changed to use `/daemon`, panel changes now reflect this.
+* Only display servers that a user is owner of or subuser of in the Admin CP rather than all servers if the user is marked as an admin.
 
 ## v0.6.0-beta.2.1 (Courageous Carniadactylus)
 ### Fixed